+++ docs/changes.rst
@@ -4,6 +4,7 @@ Change log
 Pending
 -------
 
+* Fix Server-Timing header not being overridden if it exists in header already
 * Replaced ``requirements_dev.txt`` file for ``pyproject.toml`` support with
   dependency groups.
 * Updated ReadTheDocs Python version to 3.13.